Added git service to document some learning.
Need to expose the git user's home so we can persist
the repositories that will end up here.

sshd is very particular about permissions and ownership
of ~git/.ssh.  The UID and GID of my account on my dev
box are different than those of the git user inside
the container.  Creating .ssh/ and .ssh/authorized_keys
from my dev box created a directory and file for which
sshd denies access.

Will need a different way to submit public keys from a
dev box to the git user's authorized_keys.