trying yet another base image for the git service
This one does not exit when the first connection
disconnects.

The following article described PID 1 and reaping
zombie processes, but also provided details about
how to test if your container does the right thing.
http://blog.dscpl.com.au/2015/12/issues-with-running-as-pid-1-in-docker.html

But it also pointed to https://github.com/krallin/tini
a minimal tool that just forwards signals and reaps
zombies.

So trying tini instead of container-buddy.  It
turns out 0.0.5 version of container-buddy does
not manage PID1 responsibilities.

I think it is the switch from -d to -D which allows
the container to survive the first ssh connection.
But that switch also means the sshd logs don't go
to stdout.