reorganized jenkins/Dockerfile-node to group and document the layers.
I've been blindly copying the pattern of single RUN commands chained
together with lots of &&.  Today I learned more about how the most recent
version of Docker now uses content-based addressing for the layers in an
image.  Decided to do myself a favor and let there be more layers & add a
few comments to explain the grouping.

The other notable change is that the jenkins must be a member of the users
group in order to access docker.sock.